# docs/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "3X-UI Panel API",
    "version": "3.x",
    "description": "Programmatic interface to a 3X-UI panel. Authenticate either by logging in (cookie) or with an API token from Settings → Security → API Token (Bearer). All endpoints under /panel/api/* honour both modes — an API token is a full-admin credential, so treat it like the panel password."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current panel (basePath aware)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token from Settings → Security → API Token. Send as `Authorization: Bearer <token>`."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "3x-ui",
        "description": "Session cookie set by POST /login. Browser-only."
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/panel/api/inbounds/get/{id}": {
      "get": {
        "tags": [
          "Inbounds"
        ],
        "summary": "Fetch a single inbound by numeric ID.",
        "operationId": "get_panel_api_inbounds_get_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Inbound ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/inbounds/add": {
      "post": {
        "tags": [
          "Inbounds"
        ],
        "summary": "Create a new inbound. Send the full inbound payload (protocol, port, settings, streamSettings, sniffing, remark, expiryTime, total, enable). settings, streamSettings, and sniffing may be sent as nested JSON objects (preferred) or as JSON-encoded strings (legacy).",
        "operationId": "post_panel_api_inbounds_add",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "enable": true,
                "remark": "VLESS-443",
                "listen": "",
                "port": 443,
                "protocol": "vless",
                "expiryTime": 0,
                "total": 0,
                "settings": {
                  "clients": [
                    {
                      "id": "...",
                      "email": "user1"
                    }
                  ],
                  "decryption": "none",
                  "fallbacks": []
                },
                "streamSettings": {
                  "network": "tcp",
                  "security": "reality",
                  "realitySettings": {
                    "show": false,
                    "dest": "..."
                  }
                },
                "sniffing": {
                  "enabled": true,
                  "destOverride": [
                    "http",
                    "tls"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "success": false,
                  "msg": "Port 443 is already in use"
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/inbounds/del/{id}": {
      "post": {
        "tags": [
          "Inbounds"
        ],
        "summary": "Delete an inbound by ID. Also removes its associated client stats rows.",
        "operationId": "post_panel_api_inbounds_del_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Inbound ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/inbounds/bulkDel": {
      "post": {
        "tags": [
          "Inbounds"
        ],
        "summary": "Delete many inbounds in one call. Processes the list sequentially; failures are reported per id and the rest still proceed. Restarts xray at most once.",
        "operationId": "post_panel_api_inbounds_bulkDel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "ids": [
                  1,
                  2,
                  3
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": {
                    "deleted": 2,
                    "skipped": [
                      {
                        "id": 3,
                        "reason": "..."
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/inbounds/update/{id}": {
      "post": {
        "tags": [
          "Inbounds"
        ],
        "summary": "Replace an inbound’s configuration. Body shape mirrors /add. Heavy on inbounds with thousands of clients — prefer /setEnable for enable-only flips.",
        "operationId": "post_panel_api_inbounds_update_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Inbound ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/inbounds/setEnable/{id}": {
      "post": {
        "tags": [
          "Inbounds"
        ],
        "summary": "Toggle only the enable flag without serialising the whole settings JSON. Recommended for UI switches on large inbounds.",
        "operationId": "post_panel_api_inbounds_setEnable_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Inbound ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "enable": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/inbounds/{id}/resetTraffic": {
      "post": {
        "tags": [
          "Inbounds"
        ],
        "summary": "Zero out upload + download counters for a single inbound. Does not touch per-client counters.",
        "operationId": "post_panel_api_inbounds_id_resetTraffic",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Inbound ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/inbounds/{id}/delAllClients": {
      "post": {
        "tags": [
          "Inbounds"
        ],
        "summary": "Remove every client attached to a single inbound while keeping the inbound itself. Collects emails from settings.clients[] and feeds them into the optimized bulk-delete path (runtime user removal + traffic-row cleanup + SyncInbound). Destructive and cannot be undone.",
        "operationId": "post_panel_api_inbounds_id_delAllClients",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Inbound ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": {
                    "deleted": 12
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/panel/api/inbounds/pushClientTraffics": {
      "post": {
        "tags": [
          "Inbounds"
        ],
        "summary": "Receive a master panel's aggregated per-client usage, keyed by the master's GUID. Stored in a side table used only for the UI display overlay and local quota enforcement — never folded into the local counters that masters poll, so delta accounting stays intact. Called panel-to-panel by the node traffic sync job.",
        "operationId": "post_panel_api_inbounds_pushClientTraffics",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "masterGuid": "9f6c2d-…",
                "traffics": [
                  {
                    "email": "alice",
                    "up": 1048576,
                    "down": 2097152
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true
                }
              }
            }
          }
        }
      }
    },
...
  }
}